I realised that there's only one situation in which the Markov chain "generate()" method can have no choice of "next" character: when it happens to have hit the very end of its "seed" text.
So now, the __init__ method scans the seed text twice,
exiting the second pass as soon as it sees a pre-existing
psttern.

This means we can remove a test that was needed, in
generator(), for every generated character. A small
saving per loop, but the loop runs millions of times
per second on my install.